switch to App.conf instead of config
