+++ cogs/admin.py
@@ -15,10 +15,10 @@ import requests
 from disnake.ext import commands
 from prettytable import PrettyTable
 
-import config
+from config import App
 
 cd_user = commands.BucketType.user
-logger = logging.getLogger(config.LOGGER_NAME)
+logger = logging.getLogger(App.config("LOGGER_NAME"))
 
 
 class Admin(commands.Cog):
@@ -28,7 +28,7 @@ class Admin(commands.Cog):
         """Initialize the class."""
         self.bot = bot
         self.log_path = Path(log_path)
-        self.reminders = config.REMINDERS_FILE_STREAM
+        self.reminders = App.config("REMINDERS_FILE_STREAM")
 
     # Before command invoke ----------------------------------------------------
 
@@ -42,15 +42,15 @@ class Admin(commands.Cog):
         inter: disnake.ApplicationCommandInteraction
             The interaction to possibly remove the cooldown from.
         """
-        if inter.guild and inter.guild.id != config.ID_SERVER_ADULT_CHILDREN:
+        if inter.guild and inter.guild.id != App.config("ID_SERVER_ADULT_CHILDREN"):
             return inter.application_command.reset_cooldown(inter)
 
-        if inter.author.id in config.NO_COOL_DOWN_USERS:
+        if inter.author.id in App.config("NO_COOL_DOWN_USERS"):
             return inter.application_command.reset_cooldown(inter)
 
     # Commands -----------------------------------------------------------------
 
-    @commands.cooldown(config.COOLDOWN_RATE, config.COOLDOWN_STANDARD, cd_user)
+    @commands.cooldown(App.config("COOLDOWN_RATE"), App.config("COOLDOWN_STANDARD"), cd_user)
     @commands.slash_command(name="logfile", description="get the tail of the logfile")
     @commands.default_member_permissions(administrator=True)
     async def log_tail(
@@ -86,29 +86,29 @@ class Admin(commands.Cog):
 
         for i in range(num_lines):
             num_chars += len(log_lines[-i])
-            if num_chars > config.MAX_CHARS:
+            if num_chars > App.config("MAX_CHARS"):
                 break
             tail.append(log_lines[-i])
 
         await inter.edit_original_message(f"```{' '.join(tail)}```")
 
-    @commands.cooldown(config.COOLDOWN_RATE, config.COOLDOWN_STANDARD, cd_user)
+    @commands.cooldown(App.config("COOLDOWN_RATE"), App.config("COOLDOWN_STANDARD"), cd_user)
     @commands.slash_command(name="ip", description="get the external ip address for the bot")
     @commands.default_member_permissions(administrator=True)
     async def external_ip_address(self, inter: disnake.ApplicationCommandInteraction):
         """Get the external IP of the bot."""
-        if inter.author.id != config.ID_USER_SAULTYEVIL:
+        if inter.author.id != App.config("ID_USER_SAULTYEVIL"):
             return await inter.response.send_message("You don't have permission to use this command.", ephemeral=True)
 
         ip_addr = requests.get("https://api.ipify.org").content.decode("utf-8")
         await inter.response.send_message(f"```{ip_addr}```", ephemeral=True)
 
-    @commands.cooldown(config.COOLDOWN_RATE, config.COOLDOWN_STANDARD, cd_user)
+    @commands.cooldown(App.config("COOLDOWN_RATE"), App.config("COOLDOWN_STANDARD"), cd_user)
     @commands.slash_command(name="reboot", description="restart the bot")
     async def reboot(self, inter: disnake.ApplicationCommandInteraction):
 
         """Restart the bot."""
-        if inter.author.id != config.ID_USER_SAULTYEVIL:
+        if inter.author.id != App.config("ID_USER_SAULTYEVIL"):
             return await inter.response.send_message("You don't have permission to use this command.")
 
         logger.info("restarting bot with new process")
@@ -116,7 +116,7 @@ class Admin(commands.Cog):
 
         os.execv(sys.executable, ["python"] + sys.argv)
 
-    @commands.cooldown(config.COOLDOWN_RATE, config.COOLDOWN_STANDARD, cd_user)
+    @commands.cooldown(App.config("COOLDOWN_RATE"), App.config("COOLDOWN_STANDARD"), cd_user)
     @commands.slash_command(name="show_all_reminders", description="view all the reminders")
     @commands.default_member_permissions(administrator=True)
     async def show_all_reminders(self, inter: disnake.ApplicationCommandInteraction) -> coroutine:
